% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_ash.R
\name{ash1}
\alias{ash1}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
ash1(bins, m = 5, kopt = c(2, 2), quiet = TRUE)
}
\description{
ash::ash1 modified to use message() instead of regular text output
}
\keyword{internal}
