% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_ash.R
\name{ash2}
\alias{ash2}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
ash2(bins, m = c(5, 5), kopt = c(2, 2), quiet = T)
}
\description{
ash::ash2 modified to use message() instead of regular text output
}
\keyword{internal}
