% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_arguments.R
\name{AssignArgs}
\alias{AssignArgs}
\title{Compose default arguments merged with user values}
\usage{
AssignArgs(u, d)
}
\arguments{
\item{u}{list of user argument.}

\item{d}{list of default argument.}
}
\description{
\code{AssignArgs} modifies user arguments in its parent environment.
}
\examples{

def <- list(
  a = list(x = 1, y = 2, b = c(0, 0), s = list(x = 0, y = 0)),
  r = list(
    f = list(a = list(x = "a", y = "b")),
    m = list(r = c("x", "y", "f"), n = 5)
  )
)

x <- y <- b <- s <- f <- m <- NULL
AssignArgs(NULL, def)

usr <- list(b = c(5, 5))
AssignArgs(usr, def)

usr <- list(s = 5)
AssignArgs(usr, def)

usr <- list(f = ".")
AssignArgs(usr, def)

usr <- list(f = c(x = "+", y = "-"))
AssignArgs(usr, def)

M <- matrix(1:4, 2, 2)
usr <- list(
  x = M,
  m = list(
    list(y = 1:2, f = c(x = "A")),
    list(x = 3:4, f = list(y = "B"))
  )
)
AssignArgs(usr, def)
}
\seealso{
\link{AtomicArgs},
  \link{ReferenceArgs},
  \link{ComposeArgs},
  \link{FormatVectors}
}
\keyword{internal}
