% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_arguments.R
\name{AtomicArgs}
\alias{AtomicArgs}
\title{Compose default atomic values merged with user values}
\usage{
AtomicArgs(u, a)
}
\arguments{
\item{u}{user argument values.}

\item{a}{default argument values.}
}
\value{
\code{AtomicArgs} returns an R object of the same type as \code{a}.
}
\description{
Compose default atomic values merged with user values
}
\examples{

a <- list(x = "a", y = "b", z = "c")
AtomicArgs(u = NULL, a)
AtomicArgs(u = "d", a)
AtomicArgs(u = c(y = "d"), a)
AtomicArgs(u = list(y = "d"), a)
AtomicArgs(u = c(x = "d", z = "d"), a)
AtomicArgs(u = list(x = "d", z = "d"), a)

a <- list(x = 0, y = 0, z = 0)
m <- matrix(1:4, 2, 2)
AtomicArgs(u = m, a)
AtomicArgs(u = list(x = m), a)
AtomicArgs(u = list(y = m), a)
AtomicArgs(u = list(z = m), a)
}
\keyword{internal}
