% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utilities.R
\name{AutoColorParameters}
\alias{AutoColorParameters}
\title{Quick and dirty color mapping parameters}
\usage{
AutoColorParameters(colors = NULL)
}
\arguments{
\item{colors}{vector of colors (optional).}
}
\value{
\code{AutoColorParameters} returns a ColorParameters object.
}
\description{
Quick and dirty color mapping parameters
}
\seealso{
\link{colorize},
  \link{DefineColorMap}
}
\keyword{internal}
