% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize.R
\name{autoscale}
\alias{autoscale}
\title{autoscale}
\usage{
autoscale(x, mode = NULL)
}
\arguments{
\item{x}{numeric vector.}

\item{mode}{either \code{"rank"} or \code{"01"} (default).}
}
\value{
\code{autoscale} returns a numeric vector.
}
\description{
autoscale
}
\keyword{internal}
