% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barbouille.R
\name{Barbouille}
\alias{Barbouille}
\title{Global options for Barbouille functions}
\usage{
Barbouille(...)
}
\arguments{
\item{...}{Any of the following arguments:}
}
\description{
This function sets the default value of arguments used by the main functions
of the Barbouille package.
}
\seealso{
\link{ScatterPlot},
  \link{SideBySide}
}
