% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binning2D.R
\name{Binning2D}
\alias{Binning2D}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
Binning2D(
  V,
  z = NULL,
  n = 100,
  k = 0,
  xlim = NULL,
  ylim = NULL,
  breaks = TRUE,
  safe = FALSE
)
}
\arguments{
\item{V}{matrix with two columns (1st = x, 2nd = y).}

\item{z}{Not yet implemented.}

\item{n}{number of bins (1st = x, 2nd = y).}

\item{k}{smoothing (1st = x, 2nd = y).}

\item{xlim}{numeric range.}

\item{ylim}{numeric range.}
}
\value{
When \code{breaks = FALSE}, \code{Binning2D} returns a \code{matrix}.
Otherwise it returns a \code{list} with the following elements:
\item{x}{numeric vector}
\item{y}{numeric vector}
\item{z}{numeric matrix}
}
\description{
Binning2D
}
\keyword{internal}
