% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_projections.R
\name{BivariateProjection}
\alias{BivariateProjection}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
BivariateProjection(V, stencil = c("linear", "cosine", "sigmoid"),
  spray = "uniform", fwhm = 1/2)
}
\arguments{
\item{V}{numeric matrix with two columns representing bivariate observations.}

\item{stencil}{either "linear", "cosine" or "sigmoid".
It is possible to use non-ambiguous abbreviated forms of these keywords.}

\item{spray}{defines the distribution used to generate noise values.
This can be either \code{"uniform"} (default), \code{"triangle"},
\code{"cosine"}, or \code{"normal"}.
It is possible to use non-ambiguous abbreviated forms of these keywords.}

\item{fwhm}{positive numeric value defining the Full Width at Half Maximum of the normal
distribution used when \code{spray = "normal"}.
The default value is 0.5 and produces normally distributed noise values with
a similar span as when using the triangle and cosine distributions which
both have the same FWHM of 0.5 by design.}
}
\value{
\code{BivariateProjection} returns a numeric matrix representing 2D
coordinates.
}
\description{
Bivariate projection
}
\examples{

\dontrun{

n <- 50000
V <- cbind(
  c(rnorm(n, 10, 5), rnorm(n, -10, 1)),
  c(rnorm(n, 10, 1), rnorm(n, -10, 5))
)

P <- BivariateProjection(V)
r <- BivariateDensity(P)
}
}
\seealso{
\link{UnivariateProjection},
  \link{Atomize}
}
\keyword{internal}
