% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_colors.R
\name{BlendColors}
\alias{BlendColors}
\title{Mix two colors in RGB space}
\usage{
BlendColors(x, y, gamma)
}
\arguments{
\item{x}{vector of R colors.}

\item{y}{vector of R colors.}

\item{gamma}{numeric vector with values in [0 ; 1] specifying the proportion of \code{x}
when mixing \code{x} and \code{y} colors.}
}
\value{
\code{BlendColors} returns a vector of R colors.
}
\description{
Mix two colors in RGB space
}
\examples{

n <- 50

x <- 0:(n^2 - 1) / (n^2 - 1)
y <- rep(0:(n - 1) / (n - 1), length.out = length(x))

cm1 <- DefineColorMap(thresholds = 0:2/2, colors = c("blue", "yellow", "red"))
cm2 <- DefineColorMap(thresholds = 0:2/2, colors = grey(c(1, 0, 1)))

c1 <- MakeColors(x, cm1)
c2 <- MakeColors(x, cm2)

clr <- BlendColors(c1, c2, gamma = y)

plot(x, y, col = clr, pch = 20, cex = 2)

cm1 <- AutoColorParameters("Wry")
cm2 <- AutoColorParameters("WBW")

c1 <- MakeColors(x, cm1)
c2 <- MakeColors(x, cm2)

clr <- BlendColors(c1, c2, gamma = y)

plot(x, y, col = clr, pch = 20, cex = 2)

}
\keyword{internal}
