% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlot.R
\name{BoxPlot}
\alias{BoxPlot}
\title{A visually cleaned alternative to the builtin graphics::boxplot function}
\usage{
BoxPlot(x, whiskers = c(0.05, 0.25, 0.5, 0.75, 0.95), medpch = 20,
  medcex = 0.8, ...)
}
\arguments{
\item{x}{numeric matrix or data.frame.}

\item{whiskers}{numeric vector.}

\item{medpch}{see \link{bxp}.}

\item{medcex}{see \link{bxp}.}

\item{...}{optional arguments forwarded to the \link{bxp} function.}
}
\description{
A visually cleaned alternative to the builtin graphics::boxplot function
}
