% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ColorChannel<-}
\alias{ColorChannel<-}
\title{Replace color channel values}
\usage{
ColorChannel(x, k) <- value
}
\arguments{
\item{x}{character vector of colors.}

\item{k}{name of a color channel (e.g. "red", "green", "blue" and "alpha"
for transparency). Abbreviated names are allowed (e.g. "r", "g", "b", "a").}

\item{value}{numeric values between \code{0} and \code{1}.}
}
\value{
\code{ColorChannel} returns a character vector of RGBA colors in hexadecimal.
}
\description{
Replace color channel values
}
\seealso{
\link{TransformColors}
}
\keyword{internal}
