% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ColorChannel}
\alias{ColorChannel}
\title{Extract color channel values}
\usage{
ColorChannel(x, k)
}
\arguments{
\item{x}{character vector of colors.}

\item{k}{name of a color channel (e.g. "red", "green", "blue" and "alpha"
for transparency). Abbreviated names are allowed (e.g. "r", "g", "b", "a").}
}
\value{
\code{ColorChannel} returns a numeric vector.
}
\description{
Extract color channel values
}
\seealso{
\link{TransformColors}
}
\keyword{internal}
