% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize.R
\name{colorize}
\alias{colorize}
\title{Quick and dirty color mapping}
\usage{
colorize(x, mode = "0_1", colors = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{mode}{either \code{"rank"} or \code{"0_1"} (default).}

\item{colors}{vector of colors (optional).}
}
\value{
colorize returns a vector of colors
}
\description{
Quick and dirty color mapping
}
