% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorLegend.R
\name{colorLegend}
\alias{colorLegend}
\title{Legend for color maps}
\usage{
colorLegend(pos, parameters, ticks = NULL, labels = NULL,
  resolution = 100, log = F, size = c(40, 3), margin = 5, horiz = F,
  tick.pos = 1, tick.size = 1.5, offset = 0.3, lwd = 1, border = T,
  ...)
}
\arguments{
\item{pos}{the legend location which can be specified using either a single keyword in
\code{"bottomright", "bottom", "bottomleft", "left", "center", "right",
"topleft", "top", "topright"}, or the corresponding abbreviation
(\code{"br", "b", "bl", "l", "c", "r", "tl", "t", "tr"}),
or the corresponding index
(from \code{1} for \code{"bottomright"} to \code{9} for \code{"topleft"}).}

\item{parameters}{list of color mapping parameters defined with \link{defineColors}.}

\item{ticks}{values of the tick marks for graduation of the color scale
(default = \code{thresholds} as defined in \code{parameters}).}

\item{labels}{text labels shown next to the tick marks
(default = \code{ticks}).}

\item{resolution}{the resolution of the color scale or number of elementary units composing the
represented color gradients (default = \code{100}).}

\item{log}{logical. If \code{TRUE}, the color legend is represented in log scale
(default = \code{F}).}

\item{size}{the size of the color legend given as a list or a vector in the form
\code{c(}axial length, width\code{)} which are expressed in percentage of
the dimensions of the plot area (default = \code{c(40, 3)}).}

\item{margin}{numeric vector controlling the spacing between the legend and the borders
of the plot area and expressed in percentage of its dimensions
(default = \code{c(5, 5, 5, 5)} for left, right, bottom, top).}

\item{horiz}{logical controlling the legend layout.
If \code{TRUE} the color legend spans horizontally from the given location,
and vertically otherwise (default = \code{F}, vertical).}

\item{tick.pos}{relative location of graduations, which is indicated by a value of \code{1}
or \code{-1} for a location to the right/above or to the left/below
the color scale respectively.}

\item{tick.size}{length of the tick marks, expressed in percentage of the dimensions of the
plot area (default = \code{1.5}).}

\item{offset}{spacing between tick marks and tick labels.}

\item{lwd}{line width for tick marks and color scale borders.}

\item{border}{logical activating the drawing of black borders around the color scale
(default = \code{T}).}

\item{...}{optional parameters passed to the \link{text} function.}
}
\description{
colorLegend draws a color legend according to given color mapping parameters.
}
\examples{
# Radial color gradient with two normally distributed random variables
x <- rnorm(2000)
y <- rnorm(2000)
z <- 10^sqrt(x^2 + y^2)

col.par <- defineColors(name = "bantignies")
clrs    <- makeColors(z, parameters = col.par)

plot(x, y, xlim = c(-4.5, 4.5), ylim = c(-5, 4), pch = 20, col = clrs)
colorLegend(
  "bottom", parameters = col.par, log = T, size = c(70, 3), horiz = T
)
}
\seealso{
\link{makeColors},
  \link{defineColors}
}
\author{
Benjamin Leblanc
}
