% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_arguments.R
\name{ComposeArgs}
\alias{ComposeArgs}
\title{Compose default arguments merged with user values}
\usage{
ComposeArgs(u, d)
}
\arguments{
\item{u}{list of user argument.}

\item{d}{list of default argument.}
}
\value{
\code{ComposeArgs} returns a \code{list}.
}
\description{
Compose default arguments merged with user values
}
\keyword{internal}
