% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{EmptyPlot}
\alias{EmptyPlot}
\title{As the name suggests...}
\usage{
EmptyPlot(axes = T, xlab = "", ylab = "", ...)
}
\arguments{
\item{axes}{logical, show axes (default = T, yes)}

\item{xlab}{character, name of the horizontal axis (default = none).}

\item{ylab}{character, name of the vertical axis (default = none).}

\item{...}{optional arguments forwarded to the \link{plot.default} function.}
}
\description{
As the name suggests...
}
