% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_matrixes.R
\name{ExtractSelection}
\alias{ExtractSelection}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
ExtractSelection(M, cols, rows = NULL, meta = NULL)
}
\arguments{
\item{M}{matrix.}

\item{cols}{list.}

\item{rows}{optional.}

\item{meta}{optional data.frame.}
}
\value{
\code{ExtractSelection} returns a \code{matrix}.
}
\description{
Extract and rename matrix columns.
}
\examples{

meta <- data.frame(
  antibody  = gl(2, 2, 8, labels = c("H3", "IgG")),
  genotype  = gl(2, 1, 8, labels = c("WT", "mutant")),
  replicate = gl(2, 4, 8, labels = c("R1", "R2"))
)
rownames(meta) <- apply(meta, 1, paste, collapse = "_")
M <- matrix(1:80, 10, 8, dimnames = list(NULL, rownames(meta)))
n <- ncol(M)

# Column selection by name or with logical expression on metadata
ExtractSelection(
  M, meta = meta, cols = list(
    x = antibody == "H3"  & genotype == "WT", y = c("IgG_WT_R1", "IgG_WT_R2")
  )
)
ExtractSelection(
  M, meta = meta, cols = list(
    x = c("H3_WT_R1", "H3_WT_R2"), y = antibody == "IgG" & genotype == "WT"
  )
)

# Internal order of extracted columns is determined by metadata
ExtractSelection(
  M[, n:1], meta = meta[1:n, ], cols = list(
    x = antibody == "H3"  & genotype == "WT",
    y = antibody == "IgG" & genotype == "WT"
  )
)
ExtractSelection(
  M[, 1:n], meta = meta[n:1, ], cols = list(
    x = antibody == "H3"  & genotype == "WT",
    y = antibody == "IgG" & genotype == "WT"
  )
)
}
\keyword{internal}
