% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_arguments.R
\name{FormatVectors}
\alias{FormatVectors}
\title{Format simple vector arguments}
\usage{
FormatVectors(u, n, names = NULL)
}
\arguments{
\item{u}{user values.}

\item{n}{required vector length.}

\item{names}{optional names.}
}
\value{
\code{FormatVectors} returns an R vector of the same type as \code{u}.
}
\description{
Format simple vector arguments
}
\examples{

nx <- ny <- 0
u <- list(nx = nx, ny = ny)
FormatVectors(u, 3)
}
\seealso{
\link{AtomicArgs},
  \link{ReferenceArgs},
  \link{ComposeArgs},
  \link{AssignArgs}
}
\keyword{internal}
