% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram2D.R
\name{histogram2D}
\alias{histogram2D}
\title{Bivariate histograms}
\usage{
histogram2D(x, y, nx = 100, ny = 100, xlim = NULL, ylim = NULL,
  plot = F, ...)
}
\arguments{
\item{x}{}

\item{y}{}

\item{nx}{number of bins on x values}

\item{ny}{number of bins on y values}

\item{xlim}{range of x values}

\item{ylim}{range of y values}

\item{plot}{}

\item{...}{}
}
\value{
histogram2D returns a \code{list} with the following elements:
\code{x}, \code{y}, \code{z}.
}
\description{
Bivariate histograms
}
\author{
Benjamin Leblanc
}
\seealso{
\link{plotImage},
  \link{defineColors},
  \link{scatterPlot}
}

