% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{hsv2R}
\alias{hsv2R}
\title{Convert an HSV matrix into R colors (hexadecimal)}
\usage{
hsv2R(x)
}
\arguments{
\item{x}{numeric matrix with 3 columns representing hue (H), saturation (S), and
value (V) color components respectively}
}
\value{
character vector of R colors (hexadecimal)
}
\description{
Convert an HSV matrix into R colors (hexadecimal)
}
\seealso{
\link{rgb2R},
  \link{R2hsv},
  \link{R2rgb},
  \link{rgb2hsv},
  \link{hsv2rgb}
}

