% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{hsv2rgb}
\alias{hsv2rgb}
\title{Convert colors from HSV matrix to RGB matrix}
\usage{
hsv2rgb(x)
}
\arguments{
\item{x}{numeric matrix with 3 columns representing hue (H), saturation (S), and
value (V) color components respectively}
}
\value{
numeric matrix with 3 columns representing red (R), green (G) and blue (B)
color components respectively
}
\description{
Convert colors from HSV matrix to RGB matrix
}
\seealso{
\link{rgb2hsv},
  \link{hsv2R},
  \link{rgb2R},
  \link{R2hsv},
  \link{R2rgb}
}
\keyword{internal}
