% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_matrixes.R
\name{List2DataTable}
\alias{List2DataTable}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
List2DataTable(x, key, columns, ...)
}
\arguments{
\item{x}{list containing table values.}

\item{key}{name of the key column.}

\item{columns}{column names.}
}
\description{
Convert a list into a data.table.
}
\keyword{internal}
