% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_simulations.R
\name{LocalShuffle}
\alias{LocalShuffle}
\title{Partial randomization}
\usage{
LocalShuffle(x, k, is.sorted = FALSE)
}
\arguments{
\item{x}{numeric vector.}

\item{k}{integer.}

\item{is.sorted}{logical.}
}
\description{
Partial randomization
}
\keyword{internal}
