% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeColors.R
\name{makeColors}
\alias{makeColors}
\title{Piecewise color mapping and group-associated coloring}
\usage{
makeColors(v, thresholds = NULL, colors = NULL, range = NULL,
  number = NULL, below = "white", above = "white", na = NA,
  levels = 256, name = "", parameters = NULL, alpha = 1, override = T,
  grp = NULL, grp.prm = NULL)
}
\arguments{
\item{v}{vector of numeric values.}

\item{thresholds}{numeric vector defining the boundaries for each range of values to be
mapped into a specific color interval. When unspecified, thresholds are
automatically set according to the \code{range} and the \code{colors} or
\code{number} parameters.}

\item{colors}{color vector providing the colors representing each threshold value
(default = \link{rainbow}).}

\item{range}{range of the numeric values to be represented
(default = \code{range(thresholds)}).}

\item{number}{the number of color intervals, used only if thresholds and colors
are unspecified (default = \code{2} in that case).}

\item{below}{color used for all values below the minimum threshold
(defaut = \code{"white"}).}

\item{above}{color used for all values above the maximum threshold
(defaut = \code{"white"}).}

\item{na}{color used for missing (\code{NA}) values
(defaut = \code{NA}, not visible).}

\item{levels}{integer vector controlling the number of color levels generated in each
color interval (default = \code{256}, minimum = \code{2}).}

\item{name}{recall a predefined set of color mapping parameters.}

\item{parameters}{list of color mapping parameters defined by \link{defineColors}.}

\item{alpha}{transparency (numeric value(s) between \code{0} and \code{1}).}

\item{override}{logical determining if the transparency of below and above colors should
override the provided alpha value(s) (default = T, yes).}

\item{grp}{group memberships (default = none).
See \link{groupIndex} for documentation on how to specify group memberships.}

\item{grp.prm}{data.frame of group representation parameters defined by \link{defineGroups}.}
}
\value{
makeColors returns a character vector of RGBA colors in hexadecimal.
}
\description{
makeColors builds a vector of RGB colors that may represent numerical values
or group memberships or a combination of both.
The color representation of numerical values is generated by applying
piecewise color mapping to these values, with precise control of the
underlying color scale.
Group coloring consist in using specific colors to represent predefined group
memberships. See \link{defineGroups} for documentation on group
representation parameters and \link{groupIndex} for documentation on how to
specify group memberships in a finite set of elements.
Group coloring can be used exclusively by leaving all the piecewise color
mapping parameters unspecified (which implies not providing any values for
the first argument of this function).
}
\examples{

# Radial color gradients with two normally distributed random variables

layout(matrix(1:4, 2, 2, byrow = T))

x <- rnorm(2000)
y <- rnorm(2000)
z <- sqrt(x^2 + y^2)
a <- atan2(y, x) + pi / 2

alpha <- ifelse(cos(8 * a) > 0, 0, 1)

clr.prm <- defineColors(
  c(0, 2), c("black", "white"), above = "red", range = c(0, 2.3)
)
clr <- makeColors(z, parameters = clr.prm, alpha = alpha, override = F)
plot(x, y, xlim = c(-5, 5), ylim = c(-5, 5), pch = 20, col = clr)
colorLegend("topleft", parameters = clr.prm, cex = 0.8)

clr.prm <- defineColors(
  seq(0, 2, 0.5), c("black", "lightgrey"), above = "red", range = c(0, 2.3)
)
clr <- makeColors(z, parameters = clr.prm)
plot(x, y, xlim = c(-5, 5), ylim = c(-5, 5), pch = 20, col = clr)
colorLegend("topleft", parameters = clr.prm, cex = 0.8)

clr.prm <- defineColors(
  thresholds = c(0, 2), colors = c("black", "lightgrey"), above = "red",
  range = c(0, 2.3), levels = 4
)
clr <- makeColors(z, parameters = clr.prm)
plot(x, y, xlim = c(-5, 5), ylim = c(-5, 5), pch = 20, col = clr)
colorLegend(
  "topleft", parameters = clr.prm,
  ticks = seq(0, 2, length.out = 5),
  cex = 0.8
)

# Use of a log scale in the color space
x <- runif(2000, -3, 3)
y <- runif(2000, -3, 3)
z <- 1 / sqrt(x^2 + y^2)                    # color mapped values
q <- 1/c(2, 1, 0.5)                         # thresholds
r <- exp(log(1/c(2, 0.5)) + c(-0.25, 0.25)) # range
clr.prm <- defineColors(
  q, c(grey(1:0/2), "red"), above = "lightgrey", below = "lightgrey",
  range = r
)
clr <- makeColors(z, parameters = clr.prm)
plot(x, y, xlim = c(-5, 5), ylim = c(-5, 5), pch = 20, col = clr)
colorLegend("topleft", parameters = clr.prm, cex = 0.8, log = T)
}
\author{
Benjamin Leblanc
}
\seealso{
\link{colorLegend},
  \link{defineColors},
  \link{defineGroups}
}

