% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_matrixes.R
\name{Matrix2Table}
\alias{Matrix2Table}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
Matrix2Table(X, grp = NULL)
}
\description{
Convert a multi-column numeric matrix into data.frame with two columns.
}
\keyword{internal}
