% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelHist2D.R
\name{ParallelHist2D}
\alias{ParallelHist2D}
\title{ParallelHist2D}
\usage{
ParallelHist2D(m, nx = 100, ny = 100, spacing = 0.2, jitter = "unif",
  method = "bin", plot = F, grid = T, color.mapper = NULL, ...)
}
\arguments{
\item{m}{matrix}

\item{nx}{number of bins on the x axis}

\item{ny}{number of bins on the y axis}

\item{spacing}{numeric}

\item{jitter}{either "unif" (default), "triangle", or "norm"}

\item{method}{either "bin" or "ash"}

\item{plot}{logical (default = F)}

\item{grid}{logical (default = T)}

\item{color.mapper}{function}

\item{...}{}
}
\value{
\code{ParallelHist2D} returns a \code{list} with the following elements:
\code{x}, \code{y}, \code{z}.
}
\description{
ParallelHist2D
}
