% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PieChart.R
\name{PieChart}
\alias{PieChart}
\title{A slightly improved version of the builtin graphics::pie function}
\usage{
PieChart(x, labels = names(x), edges = 200, radius = 0.8, tcl = 1.05,
  lbs = 1.1, clockwise = FALSE, init.angle = if (clockwise) 90 else 0,
  density = NULL, angle = 45, col = NULL, border = NULL, lty = NULL,
  main = NULL, ...)
}
\arguments{
\item{x}{a vector of non-negative numerical quantities.
    The values in \code{x} are displayed as the areas of pie slices.}

\item{labels}{one or more expressions or character strings giving
    names for the slices.  Other objects are coerced by
    \code{\link{as.graphicsAnnot}}.  For empty or \code{NA}
    (after coercion to character) labels, no label nor pointing line
    is drawn.}

\item{edges}{the circular outline of the pie is approximated by a
    polygon with this many edges.}

\item{radius}{the pie is drawn centered in a square box whose sides
    range from \eqn{-1} to \eqn{1}.  If the character strings labeling
    the slices are long it may be necessary to use a smaller radius.}

\item{tcl}{tick length, given as an expansion coefficient (default = 1.05).}

\item{lbs}{label spacing, given as an expansion coefficient (default = 1.1).}

\item{clockwise}{logical indicating if slices are drawn clockwise or
    counter clockwise (i.e., mathematically positive direction), the
    latter is default.}

\item{init.angle}{number specifying the \emph{starting angle} (in
    degrees) for the slices. Defaults to 0 (i.e., \sQuote{3 o'clock})
    unless \code{clockwise} is true where \code{init.angle}
    defaults to 90 (degrees), (i.e., \sQuote{12 o'clock}).}

\item{density}{the density of shading lines, in lines per inch.
    The default value of \code{NULL} means that no shading lines
    are drawn. Non-positive values of \code{density} also inhibit the
    drawing of shading lines.}

\item{angle}{the slope of shading lines, given as an angle in
    degrees (counter-clockwise).}

\item{col}{a vector of colors to be used in filling or shading
    the slices. If missing a set of 6 pastel colours is used,
    unless \code{density} is specified when \code{par("fg")} is used.}

\item{border}{(possibly vectors) arguments passed to
    \code{\link{polygon}} which draws each slice.}

\item{lty}{(possibly vectors) arguments passed to
    \code{\link{polygon}} which draws each slice.}

\item{main}{an overall title for the plot.}

\item{...}{\link{graphical parameters} can be given as arguments to
    \code{pie}.  They will affect the main title and labels only.}
}
\description{
A slightly improved version of the builtin graphics::pie function
}
\references{
\url{http://stackoverflow.com/questions/33259275/how-to-place-the-labels-further-from-pie-chart}
}
\seealso{
\link{pie}
}
