% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{PlotContour}
\alias{PlotContour}
\title{PlotContour}
\usage{
PlotContour(xy, k = 5, poly = F, ...)
}
\arguments{
\item{xy}{}

\item{k}{}

\item{poly}{}

\item{...}{}
}
\description{
PlotContour
}
\keyword{internal}
