% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PlotImage}
\alias{PlotImage}
\title{Plot a matrix of colors}
\usage{
PlotImage(m, x = NULL, y = NULL, ...)
}
\arguments{
\item{m}{matrix of color values.}

\item{x}{coordinates of the x axis bins.}

\item{y}{coordinates of the y axis bins.}
}
\description{
Plot a matrix of colors
}
