% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED_PlotSideBySide.R
\name{PlotSideBySide}
\alias{PlotSideBySide}
\alias{PlotDistributions}
\alias{PlotVariations}
\title{PlotSideBySide}
\usage{
PlotSideBySide(X, rng = NULL, grp = NULL, safe = F, bins = 200,
  uv = 50, bv = 50, mask = NULL, spray = NA, fwhm = NA,
  stencil = "linear", smoothing = NA, runs = NA, sampling = 5e+05,
  densities = "global", proportions = NULL, ordering = NULL, violin = F,
  colors = NA, gradient = "bright", saturation = 1, grp.colors = list(),
  spacing = 0, grid = grey(0.5, alpha = 0.5), axis = T, box = T,
  layout = c("horizontal", "vertical"), names = T, las = 1,
  label = NULL, ...)

PlotDistributions(...)

PlotVariations(...)
}
\arguments{
\item{X}{numeric matrix.}

\item{rng}{range.}

\item{grp}{group memberships, optional.}

\item{safe}{logical.}

\item{bins}{integer, binning of values.}

\item{uv}{integer, bins for univariate distributions.}

\item{bv}{integer, bins for bivariate distributions.}

\item{mask}{integer.}

\item{spray}{defines the distribution used to generate noise values.
This can be either \code{"uniform"} (default), \code{"triangle"},
\code{"cosine"}, or \code{"normal"}.
It is possible to use non-ambiguous abbreviated forms of these keywords.}

\item{fwhm}{positive numeric value defining the Full Width at Half Maximum of the normal
distribution used when \code{spray = "normal"}.
The default value is 0.5 and produces normally distributed noise values with
a similar span as when using the triangle and cosine distributions which
both have the same FWHM of 0.5 by design.}

\item{stencil}{either "linear", "cosine" or "sigmoid".
It is possible to use non-ambiguous abbreviated forms of these keywords.}

\item{smoothing}{vector/list, number of bins.}

\item{runs}{replicates for spraying.}

\item{sampling}{integer.}

\item{densities}{global, relative.}

\item{proportions}{static, local.}

\item{ordering}{static, local.}

\item{violin}{logical.}

\item{colors}{vector/list.}

\item{gradient}{}

\item{saturation}{coefficient.}

\item{grp.colors}{vector/list.}

\item{spacing}{integer, number of bins.}

\item{grid}{color.}

\item{box}{logical.}

\item{layout}{horizontal, vertical.}

\item{names}{logical.}

\item{las}{integer, see par for boxplot/barplot.}

\item{label}{main title.}

\item{...}{optional arguments forwarded to the \link{PlotImage} function.}
}
\description{
\code{PlotDistributions} and \code{PlotVariations} are two aliases of the
\code{PlotSideBySide} function.
}
\keyword{internal}
