% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderLayers.R
\name{PrevalenceScore}
\alias{PrevalenceScore}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
PrevalenceScore(x, y, f = c("linear", "quadratic", "glf"), e = 8)
}
\arguments{
\item{x}{numeric vector with values in [0 ; 1].}

\item{y}{numeric vector with values in [0 ; \code{x}].}
}
\value{
\code{PrevalenceScore} returns a numeric vector with values in [0 ; 1]
}
\description{
Subgroup prevalence score

Preeminence score \eqn{p} is defined as:
\deqn{ p = \sqrt{ x * ( x - y ) } }
See also the examples below.
}
\keyword{internal}
