% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{R2hsv}
\alias{R2hsv}
\title{Convert R colors into an HSV matrix}
\usage{
R2hsv(x)
}
\arguments{
\item{x}{vector of R colors}
}
\value{
numeric matrix with 3 columns representing hue (H), saturation (S), and
value (V) color components respectively
}
\description{
Convert R colors into an HSV matrix
}
\seealso{
\link{R2rgb},
  \link{hsv2R},
  \link{rgb2R},
  \link{rgb2hsv},
  \link{hsv2rgb}
}

