% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{R2rgb}
\alias{R2rgb}
\title{Convert R colors into an RGB matrix}
\usage{
R2rgb(x)
}
\arguments{
\item{x}{vector of R colors}
}
\value{
numeric matrix with 3 columns representing red (R), green (G) and blue (B)
color components respectively
}
\description{
Convert R colors into an RGB matrix
}
\seealso{
\link{R2hsv},
  \link{rgb2R},
  \link{hsv2R},
  \link{rgb2hsv},
  \link{hsv2rgb}
}
\keyword{internal}
