% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rankstat}
\alias{rankstat}
\title{Rescale x non-linearly into ]0, 1[}
\usage{
rankstat(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
\code{rankstat} returns \eqn{q = (rank(x) - 0.5) / N} where N = length(x)
}
\description{
rescale values non-linearly to the unit interval using ranks
}
\seealso{
\link{S01},
  \link{autoscale}
}
\keyword{internal}
