% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_matrixes.R
\name{ReCombine}
\alias{ReCombine}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
ReCombine(M, f = NULL)
}
\arguments{
\item{M}{numeric matrix.}

\item{f}{function.}
}
\value{
\code{ReCombine} returns a numeric \code{matrix}.
}
\description{
Combine matrix columns
}
\examples{

M <- matrix(1:40, 10, 4, dimnames = list(NULL, c("x", "x", "y", "y")))

ReCombine(M, f = c(x = "mean"))
ReCombine(M, f = c(y = "mean"))
ReCombine(M, f = list(x = "mean", y = mean))
ReCombine(M, f = list(x =  mean,  y = "mean"))
ReCombine(M, f = c(x = "merge"))
ReCombine(M, f = c(y = "merge"))
ReCombine(M, f = list(x = "merge", y = "merge"))
}
\seealso{
\link{ExtractSelection},
  \link{MetaSelect}
}
\keyword{internal}
