% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_arguments.R
\name{ReferenceArgs}
\alias{ReferenceArgs}
\title{Compose empty references merged with user values}
\usage{
ReferenceArgs(u, r)
}
\arguments{
\item{u}{user argument values.}

\item{r}{default argument names.}
}
\value{
\code{ReferenceArgs} returns a \code{list}.
}
\description{
Compose empty references merged with user values
}
\examples{

r <- letters[1:5]

ReferenceArgs(u = NULL, r)
ReferenceArgs(u = 1, r)
ReferenceArgs(u = 1:2, r)
ReferenceArgs(u = matrix(0, 2, 2), r)
ReferenceArgs(u = c(b = 1), r)
ReferenceArgs(u = list(d = 1), r)
ReferenceArgs(u = c(b = 1, d = 2), r)
ReferenceArgs(u = list(b = 1:2, d = 3:4), r)
ReferenceArgs(u = list(z = 0), r)
}
\seealso{
\link{AtomicArgs},
  \link{ComposeArgs},
  \link{AssignArgs},
  \link{FormatVectors}
}
\keyword{internal}
