% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderLayers.R
\name{RenderLayers}
\alias{RenderLayers}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
RenderLayers(LYR, master, mappers, render = c("prevalence", "maximum"),
  scoring = "linear")
}
\arguments{
\item{LYR}{numeric array.}

\item{master}{base color.}

\item{mappers}{layer colors.}

\item{render}{rendering method.}
}
\value{
color matrix.
}
\description{
RenderLayers
}
\keyword{internal}
