% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{ReplaceAlpha}
\alias{ReplaceAlpha}
\title{Replace transparency values}
\usage{
ReplaceAlpha(x, a)
}
\arguments{
\item{x}{character vector of colors.}

\item{a}{transparency given as a numeric value between \code{0} and \code{1}.}
}
\value{
\code{ReplaceAlpha} returns a character vector of RGBA colors in hexadecimal.
}
\description{
Replace transparency values
}
\seealso{
\link{transformColors}
}
