% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rgb2hsv}
\alias{rgb2hsv}
\title{Convert colors from RGB matrix to HSV matrix}
\usage{
rgb2hsv(x)
}
\arguments{
\item{x}{numeric matrix with 3 columns representing red (R), green (G) and blue (B)
color components respectively}
}
\value{
numeric matrix with 3 columns representing hue (H), saturation (S), and
value (V) color components respectively
}
\description{
Convert colors from RGB matrix to HSV matrix
}
\seealso{
\link{hsv2rgb},
  \link{rgb2R},
  \link{hsv2R}
  \link{R2rgb},
  \link{R2hsv}
}
\keyword{internal}
