% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{rgb2R}
\alias{rgb2R}
\title{Convert an RGB matrix into R colors (hexadecimal)}
\usage{
rgb2R(x)
}
\arguments{
\item{x}{numeric matrix with 3 columns representing red (R), green (G) and blue (B)
color components respectively}
}
\value{
character vector of R colors (hexadecimal)
}
\description{
Convert an RGB matrix into R colors (hexadecimal)
}
\seealso{
\link{hsv2R},
  \link{R2rgb},
  \link{R2hsv},
  \link{rgb2hsv},
  \link{hsv2rgb}
}

