% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_matrixes.R
\name{RowSampler}
\alias{RowSampler}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
RowSampler(M, min = 0, max = Inf)
}
\arguments{
\item{M}{matrix.}

\item{min}{minimum number of samples (default = 0).}

\item{max}{maximum number of samples (default = Inf).}
}
\value{
\code{RowSampler} returns an integer vector.
}
\description{
Random sampling of row indices.
}
\keyword{internal}
