% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterMaps.R
\name{ScatterMaps}
\alias{ScatterMaps}
\title{Empirical distributions as scatter plots}
\usage{
ScatterMaps(M, rng = NULL, meandiff = FALSE, safe = FALSE, x = NULL,
  y = NULL, f = NULL, xlab = NULL, ylab = NULL, maps = NULL,
  pops = NULL, layers = c("pops", "maps"), colors = NULL, main = NULL,
  ...)
}
\description{
Empirical distributions as scatter plots
}
\seealso{
\link{SideBySide}
}
