% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED_SideBySideDensity.R
\name{SideBySideDensity}
\alias{SideBySideDensity}
\title{Compact visualization of multiple empirical distributions}
\usage{
SideBySideDensity(m, nx = 25, ny = 200, method = "bin", jitter = "unif",
  spacing = 0.2, smoothx = FALSE, ash = list(m = c(3, 3)), xlim = NULL,
  ylim = NULL, plot = TRUE, mapper = NULL, parameters = NULL,
  global = FALSE, x.labels = TRUE, las = 1, grid = TRUE, ...)
}
\arguments{
\item{m}{matrix of numerical values.}

\item{nx}{number of bins on the horizontal axis (representing matrix columns).
The default value is 25 bins for each column in the matrix \code{m}.}

\item{ny}{number of bins on the vertical axis
(representing the range of matrix values).}

\item{method}{either "bin" (default) or "ash" for hard or smoothed binning
(see \link{ash2}).}

\item{jitter}{either "unif" (default), "norm" or "triangle".}

\item{spacing}{numeric, spacing factor along the horizontal axis
from 0 (no spacing) to 1 (maximal spacing).
By default \code{spacing = 0.2}.}

\item{smoothx}{smoothing factor along the horizontal axis, in number of bins
(defaut = FALSE, disabled).}

\item{ash}{list of arguments passed to the \link{ash} function when using this method.
(defaut = \code{list(m = c(3, 3))}).}

\item{xlim}{numeric range (default = NULL, automatic).}

\item{ylim}{numeric range (default = NULL, automatic).}

\item{plot}{logical (default = TRUE, yes).}

\item{mapper}{color mapping function. If \code{NULL} \code{SideBySideDensity} uses
the \link{colorize} function with \code{mode = "rank", color = NULL}.}

\item{parameters}{optional parameters passed to the color mapping function
(default = NULL, none).}

\item{global}{logical, use global color mapping (default = FALSE, per column color mapping).}

\item{x.labels}{logical, show column labels (default = TRUE).}

\item{las}{interger, controls the orientation of column labels (see \link{param}).}

\item{grid}{logical, show grid (default = TRUE).}

\item{...}{optional arguments forwarded to the \link{EmptyPlot} function.}
}
\value{
\code{SideBySideDensity} returns a \code{list} with the following elements:
\code{x}, \code{y}, \code{z}.
}
\description{
\code{SideBySideDensity} generates and optionally plots the empirical
distributions from a matrix of numerical values, representing the matrix
columns side by side.
This produces compact and detailed visualizations of empirical distributions
that can be used in place of boxplots or violin plots.
}
\seealso{
\link{BivariateDensity},
  \link{colorize},
  \link{ColorLegend},
  \link{PlotImage}
}
\keyword{internal}
