% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_simulations.R
\name{SimulateData}
\alias{SimulateData}
\title{SimulateData}
\usage{
SimulateData(p, m, n = 10000)
}
\arguments{
\item{p}{dataframe of mode generators with columns f, a, b (e.g. "rnorm", mu, sigma)}

\item{m}{matrix of observation groups (rows) where:
column  1 = population
column  2 = pairing correlation [-1 ; 1] (variable sort and shuffle)
columns 3.. path of modes (as indexes within p)}

\item{n}{population of each group}
}
\value{
\code{SimulateData} returns a \code{list}.
}
\description{
SimulateData
}
\keyword{internal}
