% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperRainbow.R
\name{SuperRainbow}
\alias{SuperRainbow}
\title{Rainbow colors}
\usage{
SuperRainbow(n, mod = "++", s.rng = 1, l.rng = 1, f = "hsv",
  alpha = 1, ordering = NULL)
}
\arguments{
\item{n}{number of colors.}

\item{mod}{cyclic modulation pattern, character vector.}

\item{s.rng}{range of modulated chroma/saturation values.}

\item{l.rng}{range of modulated luminance values.}

\item{f}{"hsv" or "hcl".}

\item{alpha}{transparency value (default = 1). See alpha argument in function \link{hcl}.}

\item{ordering}{color order}

\item{l.mod}{logical, generate colors with cyclic modulation of luminance values.
(default = T).}
}
\value{
SuperRainbow returns a character vector of RGBA colors in hexadecimal.
}
\description{
Rainbow colors
}
