% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformColors.R
\name{transformColors}
\alias{transformColors}
\title{Color transformation}
\usage{
transformColors(v, delta.H = NULL, S.range = NULL, V.range = NULL,
  R.range = NULL, G.range = NULL, B.range = NULL)
}
\arguments{
\item{v}{a standard R color vector or a colorParameters or groupParameters object
defined by \link{defineColors} or by \link{defineGroups}.}

\item{delta.H}{vector of delta hue values (default = NULL, no effect).}

\item{S.range}{the targeted saturation range (default = NULL, no effect).}

\item{V.range}{the targeted value range (default = NULL, no effect).}

\item{R.range}{targeted range for the red component (default = NULL, no effect).}

\item{G.range}{targeted range for the green component (default = NULL, no effect).}

\item{B.range}{targeted range for the blue component (default = NULL, no effect).}
}
\value{
list of color mapping parameters with transformed threshold colors.
}
\description{
transformColors modifies a vector of colors in HSV or RGB spaces. If
both HSV and RGB transformation parameters are provided, the colors will be
modified first in HSV space, then in RGB space.
When providing color or group parameters as input, transformColors returns
color mapping parameters with modified threshold colors or group
representation parameters with modified group-associated colors.
}
\examples{
layout(matrix(1:4, 2, 2))

col.prm <- defineColors(c(1,101), c("purple", "yellow"), levels = 5)
image(matrix(1:100, 10), col = makeColors(1:100, parameters = col.prm))

# Transform to grey levels
tst <- transformColors(col.prm, S.range = 0)
image(matrix(1:100, 10), col = makeColors(1:100, parameters = tst))

# Changes hues to complementary colors
tst <- transformColors(col.prm, delta.H = 180)
image(matrix(1:100, 10), col = makeColors(1:100, parameters = tst))

# Maximize constrast
tst <- transformColors(col.prm, V.range = c(0, 1))
image(matrix(1:100, 10), col = makeColors(1:100, parameters = tst))
}
\author{
Benjamin Leblanc
}
\seealso{
\link{defineColors},
  \link{defineGroups},
  \link{makeColors}
}

