% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_projections.R
\name{UnivariateProjection}
\alias{UnivariateProjection}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
UnivariateProjection(X, Y = NULL, grp = NULL, proportions = NULL,
  ordering = NULL, rounding = 2, spray = "uniform", fwhm = 1/2,
  violin = FALSE, vmode = "linear")
}
\arguments{
\item{X}{numeric vector or matrix with positive values, for instance representing
empirical densities or weights. When providing a matrix, each column of X
corresponds to a group, and the membership of each observation (rows) is
being defined by the \code{grp} argument.}

\item{grp}{vector representing group memberships, required only when the \code{X}
argument is a matrix with more than 1 column.
In this case, the length of \code{grp} must be equal to \code{nrow(X)}
and its values must be integers between 1 and \code{ncol(X)}.}

\item{proportions}{either "static", "global", "local" or NULL (default, no proportions).
In static mode, the areas representing each group are equal.
In global mode, these areas are proportional to the global population of
each group and in local mode, these areas are proportional to the local
population of each group.}

\item{ordering}{either "static", "global" or "local" or NULL  (default, no ordering).
In static mode, groups are ordered by their numeric identifier.
In global mode, groups are ordered by their global population and in local
mode, groups are ordered by their local population.}

\item{rounding}{integer (default = 2).}

\item{spray}{defines the distribution used to generate noise values.
This can be either \code{"uniform"} (default), \code{"triangle"},
\code{"cosine"}, or \code{"normal"}.
It is possible to use non-ambiguous abbreviated forms of these keywords.}

\item{fwhm}{positive numeric value defining the Full Width at Half Maximum of the normal
distribution used when \code{spray = "normal"}.
The default value is 0.5 and produces normally distributed noise values with
a similar span as when using the triangle and cosine distributions which
both have the same FWHM of 0.5 by design.}

\item{violin}{logical (default = FALSE, no).}

\item{vmode}{either "linear" (default) or "rank".}
}
\value{
\code{UnivariateProjection} returns a numeric vector representing 1D
coordinates.
}
\description{
Univariate projection
}
\seealso{
\link{BivariateProjection},
  \link{Atomize}
}
\keyword{internal}
