% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDefinition.R
\name{updateDefinition}
\alias{updateDefinition}
\title{Update color and group parameters}
\usage{
updateDefinition(prm, ...)
}
\arguments{
\item{prm}{list of color or group parameters, defined with \link{defineColors} or
\link{defineGroups} respectively.}

\item{...}{parameters to be updated.}
}
\value{
updated color or group parameters
}
\description{
Update color and group parameters
}
\seealso{
\link{defineColors},
\link{defineGroups}
}
