% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagram.R
\name{venn.layout}
\alias{venn.layout}
\title{2D layout by optimization of area proportions}
\usage{
venn.layout(areas, shape = Circles, start.layout = circular.layout,
  expand = 1.2, k = NULL, move = NULL, epsilon = NULL, imax = 1000,
  show.moves = F, show.energy = F, ...)
}
\arguments{
\item{areas}{representing areas and overlaps for a set of geometric shapes}

\item{shape}{either \code{Circles} (default) or \code{Rectangles}}

\item{start.layout}{function providing an initial position for each geometric shape, either
\code{circular.layout} (default) or \code{grid.layout}}

\item{expand}{plot area expansion factor used to provide plot margins}
}
\value{
venn.layout returns a \code{list} with the following elements:
\code{obj}, \code{xlim}, \code{ylim}, \code{iterations},
\code{delta}.
}
\description{
2D layout by optimization of area proportions
}
