% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagram.R
\name{vennDiagram}
\alias{vennDiagram}
\title{Venn diagrams with approximate area proportions}
\usage{
vennDiagram(ovl = NULL, grp = NULL, grp.prm, shape = Circles,
  start.layout = circular.layout, ...)
}
\arguments{
\item{ovl}{counts matrix representing the overlaps between different groups of elements.
Either \code{ovl} or the \code{grp} parameter should be used to define group
overlaps.}

\item{grp}{group memberships allowing to derive group overlaps. When provided, the
derived group overlap data overrides any value of the \code{ovl} parameter.}

\item{grp.prm}{data.frame of group representation parameters defined by \link{defineGroups}.}

\item{shape}{either \code{Circles} (default) or \code{Rectangles}}

\item{start.layout}{either \code{circular.layout} (default) or \code{grid.layout}}

\item{...}{optional parameters passed to the \link{venn.layout} function.}
}
\description{
Venn diagrams with approximate area proportions
}
\seealso{
\link{defineGroups},
  \link{groupLegend}
}
