% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utilities.R
\name{xylim}
\alias{xylim}
\title{Make plot limits including space for legends}
\usage{
xylim(x, y = NULL, symetric = F, spacing = 0, margin = 0)
}
\arguments{
\item{x}{numeric vector or matrix with two columns.}

\item{y}{numeric vector (default = NULL).}

\item{symetric}{logical (default = F, no)}

\item{spacing}{numeric.}

\item{margin}{numeric.}
}
\value{
\code{xylim} returns a list.
}
\description{
Make plot limits including space for legends
}
\keyword{internal}
