% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize.R
\name{AutoColorParameters}
\alias{AutoColorParameters}
\title{Quick and dirty color parameters}
\usage{
AutoColorParameters(colors = NULL)
}
\arguments{
\item{colors}{vector of colors (optional).}
}
\value{
\code{AutoColorParameters} returns a \code{list}
(S3 class = \code{colorParameters}).
}
\description{
Quick and dirty color parameters
}
\keyword{internal}
