% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utilities.R
\name{autoscale}
\alias{autoscale}
\title{As the name suggests...}
\usage{
autoscale(x, mode = NULL)
}
\arguments{
\item{x}{numeric vector.}

\item{mode}{either \code{"rank"} or \code{"01"} (default).}
}
\value{
\code{autoscale} returns a numeric vector.
}
\description{
As the name suggests...
}
\seealso{
\link{colorize},
  \link{S01},
  \link{RankScore}
}
\keyword{internal}
