% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateDensity.R
\name{BivariateDensity}
\alias{BivariateDensity}
\title{Bivariate empirical distribution}
\usage{
BivariateDensity(x, y = NULL, nx = 100, ny = nx, xlim = NULL,
  ylim = NULL, method = c("bin", "ash", "bin.table"), plot = F,
  clr.mapper = NULL, parameters = NULL, alpha = 1, xlab = NULL,
  ylab = NULL, ...)
}
\arguments{
\item{x}{numeric vector or matrix with two columns.}

\item{y}{numeric vector (default = NULL).}

\item{nx}{number of bins on x values.}

\item{ny}{number of bins on y values.}

\item{xlim}{range of x values.}

\item{ylim}{range of y values.}

\item{method}{either "bin" or "ash" for hard or smoothed binning (see \link{ash2}).}

\item{plot}{logical (default = F).}

\item{clr.mapper}{color mapping function. If \code{NULL} \code{BivariateDensity} uses
the \link{colorize} function with \code{mode = "01", color = "WB"}.}

\item{parameters}{optional list of arguments passed to the color mapping function
(default = NULL, none).}

\item{alpha}{transparency (default = 1.0, none).}

\item{...}{optional arguments forwarded to the \link{PlotImage} function.}
}
\value{
\code{BivariateDensity} returns a \code{list} with the following elements:
\code{x}, \code{y}, \code{z}.
}
\description{
Generate and optionally plot a bivariate empirical distribution.
}
\examples{
n <- 100000
xy <- matrix(rnorm(2 * n), n, 2)
xy[1:100,] <- 2
layout(matrix(1:9, 3, 3, byrow = T))
h <- BivariateDensity(xy, plot = T, method = "bin.table")
h <- BivariateDensity(xy, plot = T, method = "bin")
h <- BivariateDensity(xy, plot = T, method = "ash")
h <- BivariateDensity(xy, plot = T, method = "bin.table", parameters = list(mode = "rank"))
h <- BivariateDensity(xy, plot = T, method = "bin", parameters = list(mode = "rank"))
h <- BivariateDensity(xy, plot = T, method = "ash", parameters = list(mode = "rank"))
}
\seealso{
\link{SideBySideDensity},
  \link{colorize},
  \link{ColorLegend},
  \link{PlotImage}
}
