% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{colorize}
\alias{colorize}
\title{Quick and dirty color mapping}
\usage{
colorize(x, mode = NULL, clr.prm = NULL, ...)
}
\arguments{
\item{x}{numeric vector.}

\item{mode}{either \code{"rank"} or \code{"01"} (default).}

\item{clr.prm}{a ColorParameter object defined by \link{DefineColorMap}.}
}
\value{
\code{colorize} returns a vector of RGBA colors.
}
\description{
Quick and dirty color mapping
}
\seealso{
\link{autoscale},
  \link{AutoColorParameters},
  \link{ColorLegend}
}
