% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize.R
\name{colorize}
\alias{colorize}
\title{Quick and dirty color mapping}
\usage{
colorize(x, mode = NULL, clr.prm = NULL, ...)
}
\arguments{
\item{x}{numeric vector.}

\item{mode}{either \code{"rank"} or \code{"01"} (default).}

\item{clr.prm}{list of color mapping parameters defined by \link{defineColors}.}
}
\value{
\code{colorize} returns a vector of colors.
}
\description{
Quick and dirty color mapping
}
