% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorMapper.R
\name{ColorMapper}
\alias{ColorMapper}
\title{ColorMapper}
\usage{
ColorMapper(H, shift = 30, gradient = "bright", saturation = 1,
  contrast = 0.5, luminance = 1, greyscale = F)
}
\description{
ColorMapper
}
