% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineColors.R
\name{defineColors}
\alias{defineColors}
\title{Color map definition}
\usage{
defineColors(thresholds = NULL, colors = NULL, range = NULL,
  number = NULL, below = NA, above = NA, na = NA, levels = 256,
  centered = F, extra = NULL, name = "")
}
\arguments{
\item{thresholds}{numeric vector defining the boundaries for each range of values to be
mapped into a specific color interval. When unspecified, thresholds are
automatically set according to the \code{range} and the \code{colors} or
\code{number} parameters.}

\item{colors}{color vector providing the colors representing each threshold value
(default = \link{rainbow}).}

\item{range}{range of the numeric values to be represented
(default = \code{range(thresholds)}).}

\item{number}{the number of color intervals, used only if thresholds and colors
are unspecified (default = \code{2} in that case).}

\item{below}{color used for all values below the minimum threshold
(defaut = \code{NA}, not visible).}

\item{above}{color used for all values above the maximum threshold
(defaut = \code{NA}, not visible).}

\item{na}{color used for missing (\code{NA}) values
(defaut = \code{NA}, not visible).}

\item{levels}{integer vector controlling the number of color levels generated in each
color interval (default = \code{256}, minimum = \code{2}).}

\item{centered}{logical value.}

\item{extra}{numeric value between 1.0 and above, defining the fraction of the colorscale
used to represent below and above colors with \link{colorLegend}. The default
values are either 1.0, 1.15 or 1.2, depending how below and above colors are
specified.}

\item{name}{recall a predefined set of color mapping parameters.}
}
\value{
defineColors returns a \code{list} with the following elements:
\code{thresholds}, \code{colors}, \code{range}, \code{number},
\code{below}, \code{above}, \code{na}, \code{levels}, \code{centered},
\code{extra}, \code{name}.
}
\description{
defineColors provides a general method to define piecewise translations
of numeric values into colors. It can be used in combination with
\link{makeColors} and \link{colorLegend} to produce plots with precise
and easy to interpret color informations.
}
\examples{
# Radial color gradient with two normally distributed random variables
x <- rnorm(2000)
y <- rnorm(2000)
z <- 2 * sqrt(x^2 + y^2)

col.par <- defineColors(seq(0, 5, 1))
clrs    <- makeColors(z, parameters = col.par)

plot(x, y, xlim = c(-3, 3), ylim = c(-3, 3), pch = 20, col = clrs)
colorLegend("topright", parameters = col.par)
}
\seealso{
\link{updateDefinition},
  \link{makeColors},
  \link{colorLegend},
  \link{defineGroups}
}
