% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupIndex.R
\name{GroupIndex}
\alias{GroupIndex}
\title{Resolution of group memberships}
\usage{
GroupIndex(v, parameters)
}
\arguments{
\item{v}{group memberships, which can be specified using either a logical matrix,
a list composed of vectors defining each group, or a single vector defining
memberships directly. See the Examples section below for an illustration of
these 3 possibilities.}

\item{parameters}{a GroupParameters object defined by \link{DefineGroupStyles}.}
}
\value{
\link{GroupIndex} returns an integer vector
}
\description{
This function is used internally by \link{ScatterPlot}.
}
\examples{

# Simultaneous use of color mapping and group membership ////////////////////

layout(matrix(1:4, 2, 2, byrow = TRUE))

# Test data /////////////////////////////////////////////////////////////////
x <- runif(2000, -1, 1)
y <- runif(2000, -1, 1)
z <- sqrt(x^2 + y^2)

# Color mapping parameters for z
clr.prm <- DefineColorMap(seq(-2, 2, 0.5), grey(c(0.2, 0.8)))

# Define 2 groups: 1 = border, 2 = center ///////////////////////////////////
grp.prm <- DefineGroupStyles(
  ids = 1:2, colors = rgb(1:2/2,0,0), cex = c(NA, 1), pch = c(15, 18),
  label <- c("border", "center")
)

# 1. Group membership = vector of group ids /////////////////////////////////
grp <- 1 * (abs(z - 0.5) < 0.05) + 2 * (z < 0.15)

ScatterPlot(
  x, y, pch = 20,
  clr = x + y, clr.prm = clr.prm,
  grp = grp, grp.prm = grp.prm
)
GroupLegend("br", parameters = grp.prm, N = table(grp)[c("1", "2")])

# 2. Group membership = list of indices /////////////////////////////////////
grp <- list(
  `1` = which(abs(z - 0.5) < 0.05),
  `2` = which(z < 0.15)
)

ScatterPlot(
  x, y, pch = 20,
  clr = x + y, clr.prm = clr.prm,
  grp = grp, grp.prm = grp.prm
)
GroupLegend("br", parameters = grp.prm, N = sapply(grp, length))

# 3. Group membership = matrix of booleans //////////////////////////////////
grp <- cbind(
  `1` = abs(z - 0.5) < 0.05,
  `2` = z < 0.15
)
ScatterPlot(
  x, y, pch = 20,
  clr = x + y, clr.prm = clr.prm,
  grp = grp, grp.prm = grp.prm
)
GroupLegend("br", parameters = grp.prm, N = apply(grp, MAR = 2, FUN = sum))
}
\seealso{
\link{DefineGroupStyles}
  \link{GroupLegend}
  \link{ScatterPlot}
}
\keyword{internal}
