% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram2D.R
\name{histogram2D}
\alias{histogram2D}
\title{Bivariate histograms}
\usage{
histogram2D(x, y, nx = 100, ny = 100, xlim = NULL, ylim = NULL,
  plot = F, ...)
}
\value{
histogram2D returns a \code{list} with the following elements:
\code{x}, \code{y}, \code{z}.
}
\description{
Bivariate histograms
}
\author{
Benjamin Leblanc
}
\seealso{
\link{plotImage},
  \link{defineColors},
  \link{scatterPlot}
}

