% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Histogram2D.R
\name{Histogram2D}
\alias{Histogram2D}
\title{Bivariate histograms}
\usage{
Histogram2D(x, y = NULL, nx = 100, ny = 100, xlim = NULL, ylim = NULL,
  method = c("bin", "ash", "bin.table"), plot = F, color.mode = "01",
  colors = NULL, alpha = 1, ...)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{nx}{number of bins on x values}

\item{ny}{number of bins on y values}

\item{xlim}{range of x values}

\item{ylim}{range of y values}

\item{method}{either "bin" or "ash" for hard or smoothed binning}

\item{plot}{logical (default = F)}

\item{alpha}{}

\item{...}{}

\item{col}{}

\item{clr}{}
}
\value{
\code{Histogram2D} returns a \code{list} with the following elements:
\code{x}, \code{y}, \code{z}.
}
\description{
Bivariate histograms
}
\examples{
n <- 100000
xy <- matrix(rnorm(2 * n), n, 2)
xy[1:100,] <- 2
layout(matrix(1:9, 3, 3, byrow = T))
h <- Histogram2D(xy, plot = T, method = "bin.table")
h <- Histogram2D(xy, plot = T, method = "bin")
h <- Histogram2D(xy, plot = T, method = "ash")
h <- Histogram2D(xy, plot = T, method = "bin.table", color.mode = "rank")
h <- Histogram2D(xy, plot = T, method = "bin", color.mode = "rank")
h <- Histogram2D(xy, plot = T, method = "ash", color.mode = "rank")
}
\seealso{
\link{PlotImage},
  \link{defineColors},
  \link{ScatterPlot}
}
