% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram2D.R
\name{histogram2D}
\alias{histogram2D}
\title{Bivariate histograms}
\usage{
histogram2D(x, y = NULL, nx = 100, ny = 100, xlim = NULL, ylim = NULL,
  method = c("bin", "ash", "bin.table"), plot = F, color.mode = "0_1",
  colors = NULL, alpha = 1, ...)
}
\arguments{
\item{x}{}

\item{y}{}

\item{nx}{number of bins on x values}

\item{ny}{number of bins on y values}

\item{xlim}{range of x values}

\item{ylim}{range of y values}

\item{method}{}

\item{plot}{}

\item{alpha}{}

\item{...}{}

\item{col}{}

\item{clr}{}
}
\value{
histogram2D returns a \code{list} with the following elements:
\code{x}, \code{y}, \code{z}.
}
\description{
Bivariate histograms
}
\examples{
n <- 100000
xy <- matrix(rnorm(2 * n), n, 2)
xy[1:100,] <- 2
layout(matrix(1:9, 3, 3, byrow = T))
h <- histogram2D(xy, plot = T, method = "bin.table")
h <- histogram2D(xy, plot = T, method = "bin")
h <- histogram2D(xy, plot = T, method = "ash")
h <- histogram2D(xy, plot = T, method = "bin.table", color.mode = "rank")
h <- histogram2D(xy, plot = T, method = "bin", color.mode = "rank")
h <- histogram2D(xy, plot = T, method = "ash", color.mode = "rank")
}
\seealso{
\link{plotImage},
  \link{defineColors},
  \link{scatterPlot}
}
