% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{PlotImage}
\alias{PlotImage}
\title{Plot matrix of colors as an image}
\usage{
PlotImage(m, x = NULL, y = NULL, ...)
}
\arguments{
\item{m}{matrix of color values}

\item{x}{coordinates of the x axis bins}

\item{y}{coordinates of the y axis bins}
}
\description{
Plot matrix of colors as an image
}
