% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointLabels.R
\name{pointLabels}
\alias{pointLabels}
\title{Labels in scatterplots}
\usage{
pointLabels(x, y, centroid.shift = c(0, 0), distance = 1,
  pts.offset = 0.1, txt.offset = 0.2, labels = NULL, col = "black",
  plot = T, show.centroid = F, ...)
}
\arguments{
\item{x}{the x axis coordinate of points to be labelled.}

\item{y}{the y axis coordinate of points to be labelled.}

\item{centroid.shift}{displacement vector used to control the focus point of labels and associated
arrows.}

\item{distance}{the distance between points and corresponding labels
(e.g. arrows length).}

\item{pts.offset}{the spacing distance between points and arrow ends.}

\item{txt.offset}{the spacing distance between text labels and arrow ends.}

\item{labels}{character vector with text labels, one for each point to be annotated.}

\item{col}{vector of colors for text labels and arrows (default = \code{"black"}).}

\item{plot}{logical activating the drawing of text labels and arrows
(default = \code{T}).}

\item{show.centroid}{logical activating the visualization of the focus point of labels and
arrows (default = \code{F}).}

\item{...}{optional parameters passed to the \link{text} function.}
}
\value{
\code{pointLabels} returns a \code{list} with the following elements:
\item{labels}{
  character vector with text labels
}
\item{offset}{
  value of the text \code{offset} parameter
}
\item{pos}{
  value of the text \code{pos} parameter
}
\item{lab.x, lab.y}{
  the x and y coordinates for text labels
}
\item{pts.x, pts.y}{
  the x and y coordinates for arrows between text labels and corresponding
  points
}
\item{centroid}{
  coordinates of the centroid
}
}
\description{
pointLabels overlays text labels and arrows to annotate a specific set of
points inside a 2D scatterplot.
}
\author{
Benjamin Leblanc, Itys Comet
}
