% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rankstat}
\alias{rankstat}
\title{rankstat}
\usage{
rankstat(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
\eqn{(rank(x) - 0.5) / N} where N = length(x)
}
\description{
rank statistics
}
\keyword{internal}
