% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{replaceAlpha}
\alias{replaceAlpha}
\title{Replace transparency values}
\usage{
replaceAlpha(x, a)
}
\arguments{
\item{x}{character vector of colors.}

\item{a}{transparency given as a numeric value between \code{0} and \code{1}.}
}
\value{
replaceAlpha returns a character vector of RGBA colors in hexadecimal.
}
\seealso{
\link{transformColors}
}
