% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterMaps.R
\name{ScatterMaps}
\alias{ScatterMaps}
\title{Empirical distributions as scatter plots}
\usage{
ScatterMaps(M, rng = NULL, meandiff = FALSE, safe = FALSE, x = NULL,
  y = NULL, f = NULL, xlab = NULL, ylab = NULL, maps = NULL,
  pops = NULL, layers = c("pops", "maps"), colors = NULL, main = NULL,
  ...)
}
\arguments{
\item{M}{numeric matrix.}

\item{rng}{numeric range.}

\item{meandiff}{not yet implemented.}

\item{safe}{logical.}

\item{x}{matrix column(s).}

\item{y}{matrix column(s).}

\item{f}{combining function.}

\item{xlab}{horizontal axis label.}

\item{ylab}{vertical axis label.}

\item{maps}{list.}

\item{pops}{sub-populations.}

\item{layers}{either 'pops' or 'maps'}

\item{colors}{work in progress...}

\item{main}{title.}

\item{...}{optional parameters (see the \link{Barbouille} function).}
}
\description{
Empirical distributions as scatter plots
}
\seealso{
\link{SideBySide}
}
