% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter plot with color mapping and group representation}
\usage{
scatterPlot(x, y = NULL, clr = NULL, clr.prm = NULL, grp = NULL,
  grp.prm = NULL, density = c(NULL, "knn", "hist2D"), xlab = "",
  ylab = "", ...)
}
\arguments{
\item{x}{horizontal axis coordinate.}

\item{y}{vertical axis coordinate.}

\item{clr}{vector of numeric values passed to \link{makeColors} for color mapping
(default = none).}

\item{clr.prm}{list of color mapping parameters defined by \link{defineColors}.}

\item{grp}{group memberships for group highlighting (default = none).
See \link{groupIndex} for documentation on how to specify group memeberships.}

\item{grp.prm}{list of group representation parameters defined by \link{defineGroups}.}

\item{xlab}{label for the horizontal axis (default = description of x).}

\item{ylab}{label for the vertical axis (default = description of x).}

\item{...}{optional parameters passed to the \link{plot.default} function.}
}
\description{
Scatter plot with color mapping and group representation
}
\examples{
# Angular groups with two normally distributed random variables

layout(matrix(1:4, 2, 2, byrow = T))

x <- rnorm(2000)
y <- rnorm(2000)
z <- sqrt(x^2 + y^2)
a <- atan2(y, x)

grp <- 60 * (180/pi * a) \%/\% 60

grp.ids <- sort(unique(grp))
grp.lbl <- paste(grp.ids, "° < a < ", grp.ids + 60, "°", sep="")

grp.prm <- defineGroups(
  grp.ids, labels = grp.lbl, color = grey(c(0.2, 0.6)), pch = 1:6, cex = 0.7
)

scatterPlot(x, y, grp = grp, grp.prm = grp.prm, xlab = "x", ylab = "y")
groupLegend("tr", grp.prm, xjust = 1, title = "angle from atan2")

scatterPlot(a, z, grp = grp, grp.prm = grp.prm, xlab = "a (rad)", ylab = "z")
groupLegend("tr", grp.prm, xjust = 1, title = "angle from atan2")

scatterPlot(a, x, grp = grp, grp.prm = grp.prm, xlab = "a (rad)", ylab = "x")
groupLegend("tr", grp.prm, xjust = 1, title = "angle from atan2")

scatterPlot(a, y, grp = grp, grp.prm = grp.prm, xlab = "a (rad)", ylab = "y")
groupLegend("br", grp.prm, xjust = 1, title = "angle from atan2")
}
\author{
Benjamin Leblanc
}
\seealso{
\link{defineColors},
  \link{makeColors},
  \link{colorLegend},
  \link{defineGroups},
  \link{groupLegend}
}

