% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SideBySide.R
\name{SideBySide}
\alias{SideBySide}
\alias{Distributions}
\alias{Variations}
\title{SideBySide}
\usage{
SideBySide(M, rng = NULL, safe = FALSE, skip = NULL, pops = NULL,
  proportions = NULL, ordering = NULL, violin = FALSE, colors = NULL,
  main = NULL, ...)

Distributions(...)

Variations(...)
}
\description{
SideBySide
}
\seealso{
\link{ScatterMaps}
}
