% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SideBySide.R
\name{SideBySide}
\alias{SideBySide}
\alias{Distributions}
\alias{Variations}
\title{SideBySide}
\usage{
SideBySide(M, rng = NULL, safe = FALSE, skip = NULL, pops = NULL,
  proportions = NULL, ordering = NULL, violin = FALSE, colors = NULL,
  main = NULL, ...)

Distributions(...)

Variations(...)
}
\arguments{
\item{M}{numeric matrix.}

\item{rng}{numeric range.}

\item{safe}{logical.}

\item{skip}{column indices.}

\item{pops}{sub-populations.}

\item{proportions}{TODO: documentation}

\item{ordering}{TODO: documentation}

\item{violin}{logical.}

\item{colors}{work in progress...}

\item{main}{title.}

\item{...}{optional parameters (see the \link{Barbouille} function).}
}
\description{
SideBySide
}
\seealso{
\link{ScatterMaps}
}
