% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperRainbow.R
\name{SuperRainbow}
\alias{SuperRainbow}
\title{Presumably an improved rainbow color generator}
\usage{
SuperRainbow(n, mod = "++", s.rng = 1, l.rng = 1, f = "hsv",
  alpha = 1)
}
\arguments{
\item{n}{number of colors.}

\item{mod}{cyclic modulation pattern, a character vector, for instance c("+-", "-+").}

\item{s.rng}{range of modulated chroma/saturation values.}

\item{l.rng}{range of modulated luminance values.}

\item{f}{"hsv" or "hcl".}

\item{alpha}{transparency value (default = 1). See alpha argument in function \link{hcl}.}
}
\value{
\code{SuperRainbow} returns a character vector of RGBA colors in hexadecimal.
}
\description{
Presumably an improved rainbow color generator
}
\seealso{
\link{DefinecolorMap},
  \link{MakeColors}
}
