% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateDefinition.R
\name{UpdateDefinition}
\alias{UpdateDefinition}
\title{Update color mapping or group highlighting parameters}
\usage{
UpdateDefinition(prm, ...)
}
\arguments{
\item{prm}{a ColorParameters or GroupParameters object previously defined using
\link{DefineColorMap} or \link{DefineGroupStyles} respectively.}

\item{...}{list of parameter values replacing the existing ones (see the arguments of
ColorParameters and GroupParameters definition functions).}
}
\value{
\code{UpdateDefinition} returns the updated ColorParameters or a
GroupParameters object.
}
\description{
Update color mapping or group highlighting parameters
}
\seealso{
\link{DefineColorMap},
  \link{DefineGroupStyles}
}
