% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{xylim}
\alias{xylim}
\title{Plot limits with legend}
\usage{
xylim(x, y = NULL, symetric = F, spacing = 0, margin = 0)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}

\item{symetric}{logical}

\item{spacing}{numeric}

\item{margin}{numeric}
}
\value{
list
}
\description{
Plot limits with legend
}
\keyword{internal}
