% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{xylim}
\alias{xylim}
\title{Make plot limits including space for legends}
\usage{
xylim(x, y = NULL, symetric = F, spacing = 0, margin = 0)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}

\item{symetric}{logical}

\item{spacing}{percentage}

\item{margin}{percentage}
}
\value{
list
}
\description{
Make plot limits including space for legends
}
\keyword{internal}
